refactor(db,models): finalize schema — money as unsignedBigInteger, drop legacy columns, replace units table with enum, switch cache to Redis

+++ database/migrations/2026_08_05_000009_create_ingredient_batches_table.php
@@ -14,14 +14,14 @@ return new class extends Migration
             $table->decimal('quantity', 10, 3);
             $table->date('expiry_date')->nullable();
             $table->timestamp('received_at')->nullable();
-            $table->decimal('cost_per_unit', 10, 3)->default(0);
+            $table->unsignedBigInteger('cost_per_unit')->default(0);
             $table->integer('custom_order')->nullable();
             $table->string('status')->default('active');
             $table->string('batch_code')->nullable()->unique();
             $table->boolean('allow_expired_usage')->default(false);
             $table->decimal('initial_quantity', 10, 3)->nullable();
             $table->string('supplier_name')->nullable();
-            $table->decimal('total_cost', 15, 2)->nullable();
+            $table->unsignedBigInteger('total_cost')->nullable();
             $table->string('payment_status')->default('unpaid');
             $table->timestamps();
